adding error handlers

+++ Server/AuthService/Services/AccountService.cs
@@ -24,7 +24,7 @@ public class AccountService : IAccountService
         var result = await query.ToListAsync();
         if (result.Any())
             return _mapper.Map<List<AccountDtoGetRole>>(result);
-        return null;
+        return [];
     }
     public async Task<AccountDtoGetRole?> GetById(int accountID)
     {
@@ -32,9 +32,10 @@ public class AccountService : IAccountService
         query = query.Include(account => account.Role);
 
         var result = await query.Where(account => account.AccountID == accountID).FirstOrDefaultAsync();
-        if (result != null)
-            return _mapper.Map<AccountDtoGetRole>(result);
-        return null;
+        if (result == null)
+            throw new ArgumentNullException("User not found.");
+
+        return _mapper.Map<AccountDtoGetRole>(result);
     }
     public async Task Create(AccountDtoCreate AccountDto)
     {
@@ -44,7 +45,7 @@ public class AccountService : IAccountService
         var result = await query.Where(account => account.Email == AccountDto.Email).FirstOrDefaultAsync();
 
         if (result != null)
-            throw new ArgumentException("User already has account.");
+            throw new Exception("User already has account.");
 
         var account = _mapper.Map<Account>(AccountDto);
         await _accountRepository.Create(account);
@@ -53,26 +54,36 @@ public class AccountService : IAccountService
     public async Task Update(AccountDtoUpdate AccountDto)
     {
         if (AccountDto.AccountID <= 0)
-            throw new ArgumentException("User is invalid.");
-        var account = _mapper.Map<Account>(AccountDto);
-        await _accountRepository.Update(account);
+            throw new Exception("User id is invalid.");
+        
+        var accountItem = await GetById(AccountDto.AccountID);
+        if (accountItem != null)
+        {
+            var account = _mapper.Map<Account>(AccountDto);
+            await _accountRepository.Update(account);
+        }
     }
     public async Task Delete(AccountDtoUpdate AccountDto)
     {
         if (AccountDto.AccountID <= 0)
-            throw new ArgumentException("User is invalid.");
-        var account = _mapper.Map<Account>(AccountDto);
-        await _accountRepository.Delete(account);
+            throw new Exception("User is invalid.");
+
+        var accountItem = await GetById(AccountDto.AccountID);
+        if (accountItem != null)
+        {
+            var account = _mapper.Map<Account>(AccountDto);
+            await _accountRepository.Delete(account);
+        }
     }
 
     public async Task<AccountLoginDto> Login(string email, string password)
     {
         if (string.IsNullOrEmpty(email) || string.IsNullOrEmpty(password))
-            throw new ArgumentException("Email and Password cannot be empty.");
+            throw new Exception("Email and Password cannot be empty.");
 
         var account = await _accountRepository.GetOneByCriteria(a => a.Email == email && a.Password == password);
         if (account == null)
-            throw new ArgumentException("Incorret Email or Password. Please try again...");
+            throw new Exception("Incorret Email or Password. Please try again...");
         return _mapper.Map<AccountLoginDto>(account);
     }
 }
